# OpenAPI.yml
openapi:  3.0.3
info:
  title: PROJECT  BE13
  version: 1.0.0
  description:  Dokumentasi Project BE13
  contact:
    email: teguhbutarbutar160798@gmail.com
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/Anti-Gen/PROJCECT-BE13/1.0.0
  - description: Real Projetc servers
    url: http://34.172.81.180:80
...
paths:
...
  /classes:
    post:
      tags:
        - Class
      summary: Create Class
      description: membuat kelas baru
      security:
        - JWTAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                nama_kelas: "back-end batch 13"
                id_user: 1
                
      responses:
        '200':
          description: create class
          content:
            application/json:
              schema:
                type: object
                example:
                  status: "success"
                  message: "success add class"
                  data:
                    id : 1 
                    id_user: 1
                    nama_kelas: "back-end batch 13"
                    created_at: 2022:11:11
                    updated_at: 2022:11:11
                    
        '404':
          $ref: '#/components/responses/404'
        '500' :
          $ref: '#/components/responses/500'

    get:
      tags:
        - Class
      summary: Get Class
      description: Melihat data kelas
      security:
        - JWTAuth: []


      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: object
                example:
                  status: "success"
                  message: "success get mantee"
                  data:
                    - id : 1 
                      id_user: 2
                      nama_kelas: "back-end batch 13"
                      created_at: 2022:11:11
                      updated_at: 2022:11:12
                      
                    
                    - id : 2 
                      id_user: 3
                      nama_kelas: "back-end batch 14"
                      created_at: 2022:11:11
                      updated_at: 2022:11:12
        '404':
          $ref: '#/components/responses/404'  
  /classes/:
    get:
      tags:
        - Class
      summary: Get Class by Name
      description: Mencari data kelas dengan nama
      security:
        - JWTAuth: []
      parameters:
        - name: class_name
          description: "Class name"
          required: false
          in: query
          schema:
            type: string
            example: "back-end batch 13"

      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: object
                example:
                  status: "success"
                  message: "success get mantee"
                  data:
                    id : 1 
                    id_user: 2
                    nama_kelas: "back-end batch 13"
                    created_at: 2022:11:11
                    updated_at: 2022:11:12
                      
                  
        '404':
          $ref: '#/components/responses/404'  
  /classes/{id_class}:
    put:
      tags:
        - Class
      summary: Update Class data
      description: mengedit data kelas
      security:
        - JWTAuth: []
      parameters:
        - name: "id_class"
          description: "id_class"
          required: true
          in: path
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              example:
                nama_kelas: "back-end batch 13"
                id_user: 2
                
      responses:
        '200':
          description: mantee notes
          content:
            application/json:
              schema:
                type: object
                example:
                  status: "success"
                  message: "success add notes"
                  data:
                    id : 1 
                    id_user: 2
                    nama_kelas: "back-end batch 13"
                    created_at: 2022:11:11
                    updated_at: 2022:11:12
        '404':
          $ref: '#/components/responses/404'
        '500' :
          $ref: '#/components/responses/500'
    delete:
      tags:
        - Class
      summary: Delete Class
      description: menghapus Kelas
      security:
        - JWTAuth: []
      parameters:
        - name: "id_class"
          description: "id_class"
          required: true
          in: path
          schema:
            type: integer
            example: 1
      responses:
        '201':
          description: Delete berhasil.
          content:
            application/json:
              schema:
                type: object
                example:
                  status: "success"
                  message: "Delete success"
                  data:
                    nama_kelas: "back-end batch 13"
                    id_user:  "1"
                    
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'  
 


    
####################################
#  COMPONENTS
####################################
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "use Token"
  schemas:
...
    ResponseWithoutData:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
  responses:
    '404':
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            $ref: '#/components/schemas/ResponseWithoutData'
          example:
            status: "failed"
            message: "data not found"
    '500':
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            $ref: '#/components/schemas/ResponseWithoutData'
          example:
            status: "failed"
            message: "internal server error"